Add ColorSchemeUnit support
ColorSchemeUnit is a testing framework for Sublime Text color schemes:

https://github.com/gerardroche/sublime-color-scheme-unit

Example usage:

https://github.com/gerardroche/sublime-monokai-free/blob/master/.travis.yml

There are several issues of note about the implementation:

* The bootstrap must be call with the option `--with-color-scheme-unit`
option. The reason for this is to avoid unnecessary download of the
plugin for test suites that don't test any color schemes.

* Several log message format issues have been fixed.

* An edge-case issue where all arguments are not passed to the
`run_tests.py` script has been fixed.

Closes #59