Use ColorSchemeunit new run with package API
Re: https://github.com/gerardroche/sublime-color-scheme-unit/issues/18

The new API allows to run with package name so Unittesting doesn't need
to open any files before running the tests.

I also added an async parameter which allows to run the tests in non
async mode and get the result of the tests, True for pass and False for
fail. This allows UnitTesting to handle printing the OK, FAILED and DONE
messsages.