Deterministic timing for DeferrableTestCase
Sublime's worker thread is a task queue. If you put
tasks on it, they will run *ordered*.

We can use this to make the `yield` timing deterministic.
If we just `yield` (without an argument) we enqueue a task
to be run ASAP. That means it will wait for tasks in the
queue and then yield back to the test code.

E.g. from the test we invoke the unit-under-test which in turn
puts task (A) and (B) on the worker. We then `yield` which puts
itself as (C) on the queue and waits for it. Sublime will now run
(A), (B), and finally (C) ordered. The test runner now yields
control back to the test.

If (A) in turn enqueued a task (D) on the queue, it will not have
run yet, and in fact will not run at all if we not `yield` again.

See the test included here.

Using this, I could bring down the execution time of the
test in https://github.com/divmain/GitSavvy/pull/1056 from ~7.7s
to 2.7s. I could replace all `yield <condition>` expressions with
single or double `yield`s, making the test more reasonable. Bc
I know that a command e.g. 'hops' two times, I know exactly that
I have to `yield` two times.

This comes with the big caveat that coverage doesn't work on the
worker thread. This is already true right now, see
https://github.com/divmain/GitSavvy/pull/1056#issuecomment-442384812
and a big annoyance for writing functional tests. E.g. as soon
as we just issue a command which itself is implemented to run
'async' (t.i. on the worker thread) and then just `yield <condition>`
to wait for it, the code under test will not be marked as 'covered'.