Swap `set_timeout_async` to do `set_timeout`
In the first series of commits we tried to run all `yield` blocks,
and by that a lot of the code under test, on the worker thread. This
gives us more predictable timing bc a developer can count how many
times their code 'hops' and then yield that many times.

Generally, the FIFO principle is also true for the main thread. However,
choosing the main thread is a bad option. (a) Usually the code under tests
defers some work to the worker thread, and we really want to await that
work. (b) The main thread runs faster. We choosed the slower thread bc when
the worker yields a task, the main thread is usually done.

There is one major caveat when we run on the worker thread t.i. coverage
doesn't work anymore.

This commit introduces a switch which will *globally* for the time of the tests
swap `set_timeout_async` with `set_timeout`. By doing that

(a) everything runs super fast,
(b) every scheduled work/task runs on one thread, the main thread, so the FIFO
		principle holds and ensures deterministic timing,
(c) all code under test gets coverage.

This comes with one known caveat: Sublime puts '_async' events on the worker
thread on its own, t.i. it doesn't use the patched API. For example, you cannot
await such an event by just `yield`ing. You have to use `yield <delay>` or
`yield <condition>` as before.