Fix reloader after PC update
Since the Package Control update to v4 our reloader is broken.  E.g.
`PackageManager()._is_dependency(pkg_name)` throws as it's not there
anymore.  (Note the leading `_` telling us to not use this method.)

For now, copy the reloader from `GitSavvy` as it seems "newer" and works
at least locally.